requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
mysqlclient==2.1.1